+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add the `none` value to the json DSL
+- Add `is_some` and `is_none` functions to test for none values
 
 ## [0.6.0] - 2022-02-18